Move OpenAPI helpers to public headers
