+++ include/ccf/endpoint.h
@@ -4,7 +4,7 @@
 
 #include "ccf/ds/json.h"
 #include "ccf/endpoint_context.h"
-#include "ds/openapi.h"
+#include "ccf/ds/openapi.h"
 #include "kv/serialise_entry_blit.h"
 #include "service/map.h"
