+++ src/http/http_accept.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ccf/ds/nonstd.h"
-#include "http/http_status.h"
+#include "ccf/http_status.h"
 #include "node/rpc/error.h"
 #include "node/rpc/rpc_exception.h"
